move registrycli type to its file
