+++ src/cli/facade/preview.ts
@@ -2,7 +2,7 @@ import open from 'open';
 import { fromPromise } from 'universalify';
 
 import strings from '../../resources/index';
-import { RegistryCli } from '../../types';
+import { RegistryCli } from '../domain/registry';
 import { Logger } from '../logger';
 
 const preview = ({
