+++ src/cli/facade/registry-add.ts
@@ -1,6 +1,6 @@
 import { fromPromise } from 'universalify';
 import strings from '../../resources/index';
-import { RegistryCli } from '../../types';
+import { RegistryCli } from '../domain/registry';
 import { Logger } from '../logger';
 
 const registryAdd = ({
