+++ src/cli/facade/registry-ls.ts
@@ -1,7 +1,7 @@
 import { fromPromise } from 'universalify';
 
 import strings from '../../resources/index';
-import { RegistryCli } from '../../types';
+import { RegistryCli } from '../domain/registry';
 import { Logger } from '../logger';
 
 const registryLs = ({
